Update to v098r19 release.
byuu says:

Changelog:
- added nall/bit-field.hpp
- updated all CPU cores (sans LR35902 due to some complexities) to use
  BitFields instead of bools
- updated as many CPU cores as I could to use BitFields instead of union {
  struct { uint8_t ... }; }; pairs

The speed changes are mostly a wash for this. In some instances,
I noticed a ~2-3% speedup (eg SNES emulation), and in others a 2-3%
slowdown (eg Famicom emulation.) It's within the margin of error, so
it's safe to say it has no impact.

This does give us a lot of new useful things, however:

- no more manual reconstruction of flag values from lots of left shifts
  and ORs
- no more manual deconstruction of flag values from lots of ANDs
- ability to get completely free aliases to flag groups (eg GSU can
  provide alt2, alt1 and also alt (which is alt2,alt1 combined)
- removes the need for the nasty order_lsbN macro hack (eventually will
  make higan 100% endian independent)
- saves us from insane compilers that try and do nasty things with
  alignment on union-structs
- saves us from insane compilers that try to store bit-field bits in
  reverse order
- will allow some really novel new use cases (I'm planning an
  instant-decode ARM opcode function, for instance.)
- reduces code size (we can serialize flag registers in one line instead
  of one for each flag)

However, I probably won't use it for super critical code that's constantly
reading out register values (eg PPU MMIO registers.) I think there we
would end up with a performance penalty.

+++ higan/processor/arm/instructions-arm.cpp
@@ -27,7 +27,7 @@ auto ARM::arm_opcode(uint32 rm) {
 
   if(exceptionMode() && d == 15 && save) {
     cpsr() = spsr();
-    processor.setMode((Processor::Mode)cpsr().m);
+    processor.setMode((Processor::Mode)(uint)cpsr().m);
   }
 }
 
@@ -48,7 +48,7 @@ auto ARM::arm_move_to_status(uint32 rm) {
       psr.f = rm & 0x00000040;
       psr.t = rm & 0x00000020;
       psr.m = rm & 0x0000001f;
-      if(source == 0) processor.setMode((Processor::Mode)psr.m);
+      if(source == 0) processor.setMode((Processor::Mode)(uint)psr.m);
     }
   }
 
@@ -548,7 +548,7 @@ auto ARM::arm_op_move_multiple() {
     if(s && (list & 0x8000)) {
       if(mode() != Processor::Mode::USR && mode() != Processor::Mode::SYS) {
         cpsr() = spsr();
-        processor.setMode((Processor::Mode)cpsr().m);
+        processor.setMode((Processor::Mode)(uint)cpsr().m);
       }
     }
   } else {
